Refactoring of further int values, mostly converting them to size_t.
- Switched most symbolic signed constants to unsigned variants. Redefined EOF in particular to become (size)-1, to avoid having to use signed token type values.
- Introduced INVALID_INDEX for all previous -1 values to indicate e.g. not found indexes etc.
- Added 2 helpers to convert between symbolic and numeric form (mostly for intervals and toString()).
- Removed many no longer needed type casts to size_t.
- Updated templates for these changes.
- Limited runtime tests to C++ tests only, to see how Travis CI copes with that.

+++ runtime/Cpp/runtime/src/ANTLRErrorListener.h
@@ -77,7 +77,7 @@ namespace antlr4 {
     ///        the parser was able to recover in line without exiting the
     ///        surrounding rule. </param>
     virtual void syntaxError(IRecognizer *recognizer, Token *offendingSymbol, size_t line,
-                             int charPositionInLine, const std::string &msg, std::exception_ptr e) = 0;
+                             size_t charPositionInLine, const std::string &msg, std::exception_ptr e) = 0;
 
     /**
      * This method is called by the parser when a full-context prediction
@@ -187,7 +187,7 @@ namespace antlr4 {
      * was determined
      */
     virtual void reportContextSensitivity(Parser *recognizer, const dfa::DFA &dfa, size_t startIndex, size_t stopIndex,
-      int prediction, atn::ATNConfigSet *configs) = 0;
+      size_t prediction, atn::ATNConfigSet *configs) = 0;
   };
 
 } // namespace antlr4
