+++ runtime/Cpp/runtime/src/DefaultErrorStrategy.h
@@ -382,7 +382,7 @@ namespace antlr4 {
 
     virtual std::string getSymbolText(Token *symbol);
 
-    virtual int getSymbolType(Token *symbol);
+    virtual size_t getSymbolType(Token *symbol);
 
     virtual std::string escapeWSAndQuote(const std::string &s) const;
